backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Snowflake
snowflake-connector-python==3.10.0
snowflake-sqlalchemy==1.5.3
snowflake-snowpark-python==1.16.0

# Cache / Queue
redis==5.0.4
celery==5.4.0
kombu==5.3.5

# Scheduling
apscheduler==3.10.4

# HTTP client
httpx==0.27.0
aiohttp==3.9.5

# Data processing
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
python-docx==1.1.0

# Report generation
jinja2==3.1.4
weasyprint==62.3
python-pptx==0.6.23

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# AI / LLM
anthropic==0.28.0
openai==1.30.1
tiktoken==0.7.0

# File processing
python-magic==0.4.27
pypdf==4.2.0
pillow==10.3.0

# Observability
structlog==24.1.0
prometheus-client==0.20.0
opentelemetry-api==1.24.0

# Utils
python-dotenv==1.0.1
tenacity==8.3.0
pytz==2024.1
arrow==1.3.0
